Use uint

+++ main.go
@@ -23,28 +23,28 @@ type Cache struct {
 }
 
 // BinomialCoefficient is the binomial coeffcient
-func BinomialCoefficient(cache *[256][256]Cache, n, k float64) float64 {
+func BinomialCoefficient(cache *[256][256]Cache, n, k uint) float64 {
 	if k > n {
 		return 0
 	} else if k == 0 || k == n {
 		return 1
 	}
 	x := 0.0
-	if cache[int(n-1)][int(k-1)].Valid {
+	if cache[n-1][k-1].Valid {
 		x = cache[int(n-1)][int(k-1)].Value
 	} else {
 		x = BinomialCoefficient(cache, n-1, k-1)
-		cache[int(n-1)][int(k-1)].Value = x
-		cache[int(n-1)][int(k-1)].Valid = true
+		cache[n-1][k-1].Value = x
+		cache[n-1][k-1].Valid = true
 
 	}
 	y := 0.0
-	if cache[int(n-1)][int(k)].Valid {
+	if cache[n-1][k].Valid {
 		y = cache[int(n-1)][int(k)].Value
 	} else {
 		y = BinomialCoefficient(cache, n-1, k)
-		cache[int(n-1)][int(k)].Value = y
-		cache[int(n-1)][int(k)].Valid = true
+		cache[n-1][k].Value = y
+		cache[n-1][k].Valid = true
 	}
 	return x + y
 }
@@ -65,9 +65,9 @@ func Gray(input image.Image) *image.Gray16 {
 
 func main() {
 	binomial := [256][256]Cache{}
-	for n := 0; n < 256; n++ {
-		for k := 0; k <= n; k++ {
-			binomial[n][k].Value = BinomialCoefficient(&binomial, float64(n), float64(k))
+	for n := uint(0); n < 256; n++ {
+		for k := uint(0); k <= n; k++ {
+			binomial[n][k].Value = BinomialCoefficient(&binomial, n, k)
 			binomial[n][k].Valid = true
 		}
 	}
